+++ src/app/core/models/electron-api.d.ts
@@ -12,6 +12,7 @@ declare global {
       hideNotification: () => Promise<boolean>;
       hideMenu: () => Promise<boolean>;
       showMenu: () => Promise<boolean>;
+      updateMenuWithUserData?: () => Promise<boolean>;
       onMenuGenerateDaily: (cb: () => void) => void;
       onMenuProfile: (cb: () => void) => void;
       onMenuLogout: (cb: () => void) => void;